requirements.txt
Django==3.1.3
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.2
uWSGI==2.0.19.1
whitenoise==5.2.0